+++ src/soho/checkbox/soho-checkbox.component.ts
@@ -46,15 +46,12 @@ export class SohoCheckBoxComponent implements AfterViewInit, OnDestroy {
   // -------------------------------------------
   // Component Output
   // -------------------------------------------
-  /**
-   * Called when the checkbox value changes
-   */
-  @Output() onChange = new EventEmitter<SohoCheckBoxEvent>();
 
-  /**
-   * Called when the checkbox updates in some way
-   */
-  @Output() onUpdated = new EventEmitter<SohoCheckBoxEvent>();
+  /** Called when the checkbox value changes. */
+  @Output() change = new EventEmitter<SohoCheckBoxEvent>();
+
+  /** Called when the checkbox updates in some way. */
+  @Output() updated = new EventEmitter<SohoCheckBoxEvent>();
 
   /**
    * Local variables
@@ -75,8 +72,9 @@ export class SohoCheckBoxComponent implements AfterViewInit, OnDestroy {
     /**
      * Bind to jQueryElement's events
      */
-    this.jQueryElement.on('change', (e: any, args: SohoCheckBoxEvent) => this.onChange.next(args));
-    this.jQueryElement.on('updated', (e: any, args: SohoCheckBoxEvent) => this.onUpdated.next(args));
+    this.jQueryElement
+      .on('change', (e: any, args: SohoCheckBoxEvent) => this.change.next(args))
+      .on('updated', (e: any, args: SohoCheckBoxEvent) => this.updated.next(args));
 
     // no control initializer for checkbox
   }
@@ -85,7 +83,6 @@ export class SohoCheckBoxComponent implements AfterViewInit, OnDestroy {
    * Destructor.
    */
   ngOnDestroy() {
-    // No jquery control.
+    // No jQuery control.
   }
-
 }